Loader class created, this class read the ini files
