+++ keneth/shallot/config/topology.ini
@@ -1,0 +1,10 @@
+[topology]
+127.0.0.1:5000 127.0.0.1:5001
+127.0.0.1:5001 127.0.0.1:5000 127.0.0.1:5002 127.0.0.1:5003
+127.0.0.1:5002 127.0.0.1:5001 127.0.0.1:5005
+127.0.0.1:5003 127.0.0.1:5001 127.0.0.1:5004 127.0.0.1:5006
+127.0.0.1:5004 127.0.0.1:5001 127.0.0.1:5003 127.0.0.1:5007
+127.0.0.1:5005 127.0.0.1:5002 127.0.0.1:5007
+127.0.0.1:5006 127.0.0.1:5003 127.0.0.1:5006 127.0.0.1:5010
+127.0.0.1:5007 127.0.0.1:5004 127.0.0.1:5005 127.0.0.1:5010
+127.0.0.1:5010 127.0.0.1:5006 127.0.0.1:5007
