+++ keneth/shallot/data_from_ini_files_example.json
@@ -1,0 +1,96 @@
+{
+    "127.0.0.1:5000": {
+        "name": "alice",
+        "host": "127.0.0.1",
+        "port": 5000,
+        "id": 0,
+        "neighbors": [
+            "127.0.0.1:5001"
+        ]
+    },
+    "127.0.0.1:5010": {
+        "name": "bob",
+        "host": "127.0.0.1",
+        "port": 5010,
+        "id": 0,
+        "neighbors": [
+            "127.0.0.1:5006",
+            "127.0.0.1:5007"
+        ]
+    },
+    "127.0.0.1:5001": {
+        "name": "relay_1",
+        "host": "127.0.0.1",
+        "port": 5001,
+        "id": 1,
+        "neighbors": [
+            "127.0.0.1:5000",
+            "127.0.0.1:5002",
+            "127.0.0.1:5003"
+        ]
+    },
+    "127.0.0.1:5002": {
+        "name": "relay_2",
+        "host": "127.0.0.1",
+        "port": 5002,
+        "id": 2,
+        "neighbors": [
+            "127.0.0.1:5001",
+            "127.0.0.1:5005"
+        ]
+    },
+    "127.0.0.1:5003": {
+        "name": "relay_3",
+        "host": "127.0.0.1",
+        "port": 5003,
+        "id": 3,
+        "neighbors": [
+            "127.0.0.1:5001",
+            "127.0.0.1:5004",
+            "127.0.0.1:5006"
+        ]
+    },
+    "127.0.0.1:5004": {
+        "name": "relay_4",
+        "host": "127.0.0.1",
+        "port": 5004,
+        "id": 4,
+        "neighbors": [
+            "127.0.0.1:5001",
+            "127.0.0.1:5003",
+            "127.0.0.1:5007"
+        ]
+    },
+    "127.0.0.1:5005": {
+        "name": "relay_5",
+        "host": "127.0.0.1",
+        "port": 5005,
+        "id": 5,
+        "neighbors": [
+            "127.0.0.1:5002",
+            "127.0.0.1:5007"
+        ]
+    },
+    "127.0.0.1:5006": {
+        "name": "relay_6",
+        "host": "127.0.0.1",
+        "port": 5006,
+        "id": 6,
+        "neighbors": [
+            "127.0.0.1:5003",
+            "127.0.0.1:5006",
+            "127.0.0.1:5010"
+        ]
+    },
+    "127.0.0.1:5007": {
+        "name": "relay_7",
+        "host": "127.0.0.1",
+        "port": 5007,
+        "id": 7,
+        "neighbors": [
+            "127.0.0.1:5004",
+            "127.0.0.1:5005",
+            "127.0.0.1:5010"
+        ]
+    }
+}
