+++ keneth/shallot/src/classes/MessageFactory.py
